Port Mordor to visual studio 2013
Similar to upgrade on tethys:
https://gerrit.mozy.lab.emc.com/#/c/110304/

-Replace build.bat with version based on Tethys.
It will compile both 32 and 64 bit (unit tests are run against 64bit)

-Update unit tests, examples and other vcxproj files to vs2013
Fix linking, add thirdparty...props files, etc

-disable lzma unit test for windows because lzma support is not
compiled on windows

-Copy installprep project from tethys and simplify the script.
This copies openssl and other dll to build directory so that
unit test can without without requiring any special PATH.
It runs in the context of visual studio so that it can
find the dlls based on the mordor/thirdparty...props files.
As follow up cleanup tethys project can be updated to use this version.

-Warning fix

-Use the existing "v12" versions of mordor / mordortest / mordorprotobuf
from inside sln.  Eventually those will be renamed along with updates
to all windows clients.

Reviewed-on: https://gerrit.mozy.lab.emc.com/110446

+++ mordor/tests/hmac.cpp
@@ -26,9 +26,9 @@ MORDOR_UNITTEST(HMAC, HMAC_MD5_2)
 MORDOR_UNITTEST(HMAC, HMAC_MD5_3)
 {
     std::string key;
-    key.append(16, 0xaa);
+    key.append(16, (uint8_t)0xaa);
     std::string data;
-    data.append(50, 0xdd);
+    data.append(50, (uint8_t)0xdd);
     MORDOR_TEST_ASSERT(hmacMd5(data, key) ==
         "\x56\xbe\x34\x52\x1d\x14\x4c\x88\xdb\xb8\xc7\x33\xf0\xe8\xb3\xf6");
 }
